Standardize on one recipe layout
